Features/day7

+++ AdventOfCode/2019/Intcode/Argument.cs
@@ -4,22 +4,22 @@ namespace AdventOfCode._2019.Intcode
 {
     public class Argument
     {
-        private readonly IList<int> _code;
+        private readonly IList<long> _code;
         private readonly bool _immediate;
         private readonly int _position;
 
-        public Argument(IList<int> code, bool immediate, int position)
+        public Argument(IList<long> code, bool immediate, int position)
         {
             _code = code;
             _immediate = immediate;
             _position = position;
         }
 
-        public int Value => _immediate ? _code[_position] : _code[_code[_position]];
+        public long Value => _immediate ? _code[_position] : _code[(int) _code[_position]];
 
-        public void Set(int value)
+        public void Set(long value)
         {
-            _code[_code[_position]] = value;
+            _code[(int) _code[_position]] = value;
         }
     }
 }
