+++ AdventOfCode/2019/Intcode/Operations/JumpIfFalseOperation.cs
@@ -23,7 +23,7 @@ namespace AdventOfCode._2019.Intcode.Operations
             }
             else
             {
-                _program.Pointer = _arg2.Value;
+                _program.Pointer = (int) _arg2.Value;
             }
         }
     }
